ported to VS2013 & .Net 4.5 + upgraded sqlite (1.0.93, 3.8.5)

+++ SQLiteMemfs/memfs_file_data.cpp
@@ -14,7 +14,7 @@ void clear_file(memfs_file_data* pData);
 
 int create_file_data(memfs_file_data* pData, const char* zName)
 {
-    int nName;
+    size_t nName;
     memset(pData, 0, sizeof(memfs_file_data));
 
     nName = strlen(zName) + 1;
